Massive changes, to many to enumerate.  Updated based on this:
A Security is just the thing we invest in, like a stock or an option.

A Transaction is a purchase or a sale of a Security.

A Position is an opening transaction and a closing transaction.
- A position is considered open if the quantity is non-zero
- When a position is closed, the amount (realized) is just the sum of the transactions
- when a position is open, the amount is the sum of the transactions (realized) and the current price of the Security x the quantity (unrealized)

A Strategy is a collection of positions
- For now, the collection is all positions whose underlying symbol is the same

An Account is a collection of positions
- Has a balance
- Has a security where cash is kept
- Each position in an account is unique